Renaming project to 'FastWS'
