+++ .lycheeignore
@@ -4,3 +4,4 @@ https://uel.java.net/
 https://www.openapis.org/
 https://jcp.org/*
 http://localhost:8080/*
+https://github.com/QubitPi/fast-ws/blob/master/FastWS-Project-intellij-code-style.xml
