+++ docs/blog/authors.yml
@@ -1,5 +1,5 @@
 jiaqi:
   name: Jack
-  title: Maintainer of Jersey Webservice Template
+  title: Maintainer of FastWS
   url: https://github.com/QubitPi
   image_url: https://avatars.githubusercontent.com/u/16126939?v=4
