+++ docs/docs/configuration.mdx
@@ -6,7 +6,7 @@ title: Configuration
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 
-JAX-RS Jersey doesn't specify how application configurations are to be loaded. jersey-webservice-template, however,
+JAX-RS Jersey doesn't specify how application configurations are to be loaded. FastWS, however,
 utilizes a library called [owner]. 
 
 The configurations in this page can be set from several sources in the following order:
