+++ docs/docs/development.md
@@ -9,7 +9,7 @@ Template.
 Preparing for Local Development
 -------------------------------
 
-This section discusses the one-time setup in order to develop Jersey Webservice Template.
+This section discusses the one-time setup in order to develop FastWS.
 
 ### Installing Java & Maven (on Mac)
 
@@ -38,7 +38,7 @@ Make sure to execute the `sudo ln -sfn`, `echo 'export PATH=...`, and the `expor
 
 :::tip
 
-JWT is built using maven. Maven uses a separate JDK version, which can be seen via `mvn -v`. If it's not JDK 17, we
+FastWS is built using maven. Maven uses a separate JDK version, which can be seen via `mvn -v`. If it's not JDK 17, we
 should have Maven point to our JDK 17 using [JAVA_HOME](https://stackoverflow.com/a/2503679):
 
 ```bash
@@ -61,25 +61,24 @@ OpenJDK 64-Bit Server VM (build 17.0.10+9, mixed mode)
 
 ### Installing Docker Engine
 
-Jersey Webservice Template has [Docker-based integration tests]; it also supports
-[running template webserivce in Docker][jersey-webservice-template Dockerfile]. Docker can be installed by following its
+FastWS has [Docker-based integration tests]; it also supports
+[running template webserivce in Docker][fast-ws Dockerfile]. Docker can be installed by following its
 [official instructions](https://docs.docker.com/desktop/install/mac-install/)
 
 Getting Source Code
 -------------------
 
 ```bash
-git clone git@github.com:QubitPi/jersey-webservice-template.git
-cd jersey-webservice-template
+git clone git@github.com:QubitPi/fast-ws.git
+cd fast-ws
 ```
 
-### Syncing up with jersey-webservice-template's Code Styles with IntelliJ
+### Syncing up with FastWS's Code Styles with IntelliJ
 
-For the moment, we have distilled the most important code style conventions with respect to
-jersey-webservice-template's code as IntelliJ settings. If IntelliJ is used for IDE, we may import these code style
-settings by importing the [jersey-webservice-template-Project-intellij-code-style.xml][style config] file in the root
-of the repo. The setting for the project will appear as a new Scheme named "jersey-webservice-template-Project" under
-IDE's __Editor__ -> __Code Style__ section.
+For the moment, we have distilled the most important code style conventions with respect to FastWS's code as IntelliJ
+settings. If IntelliJ is used for IDE, we may import these code style settings by importing the
+[FastWS-Project-intellij-code-style.xml][style config] file in the root of the repo. The setting for the project will
+appear as a new Scheme named "fast-ws-Project" under IDE's __Editor__ -> __Code Style__ section.
 
 Please also enable "remove unused imports" by __Editor__ -> __General__ -> __Auto Import__ -> __Optimize Imports on the
 Fly__, which will automatically remove unused imports.
@@ -106,7 +105,7 @@ For IT tests, we use [Testcontainers] instead of [jcabi-mysql] because the latte
 Packaging
 ---------
 
-A [WAR file](https://en.wikipedia.org/wiki/WAR_(file_format)) named __jersey-webservice-template-1.0-SNAPSHOT.war__ will
+A [WAR file](https://en.wikipedia.org/wiki/WAR_(file_format)) named __fast-ws-1.0-SNAPSHOT.war__ will
 be generated under _target_ directory for [running in Jetty](#running-webservice-in-standalone-jetty-production) with
 the command below:
 
@@ -127,10 +126,10 @@ We can obtain the image in one of the 2 approaches below:
 
 #### Docker Hub
 
-We can pull the image from [my docker hub](https://hub.docker.com/r/jack20191124/jersey-webservice-template/):
+We can pull the image from [my docker hub](https://hub.docker.com/r/jack20191124/fast-ws/):
 
 ```bash
-docker pull jack20191124/jersey-webservice-template
+docker pull jack20191124/fast-ws
 ```
 
 #### GitHub
@@ -138,12 +137,12 @@ docker pull jack20191124/jersey-webservice-template
 We could also build the image from [source][Docker]:
 
 ```bash
-git clone https://github.com/QubitPi/jersey-webservice-template.git
-cd jersey-webservice-template
-docker build -t jack20191124/jersey-webservice-template .
+git clone https://github.com/QubitPi/fast-ws.git
+cd fast-ws
+docker build -t jack20191124/fast-ws .
 ```
 
-Please note that the `jack20191124/jersey-webservice-template` in the last command is the image name; we could replace
+Please note that the `jack20191124/fast-ws` in the last command is the image name; we could replace
 that value with anything preferred
 
 ### Standing up a Container
@@ -151,10 +150,10 @@ that value with anything preferred
 When image is built, we can spin up an instance with
 
 ```bash
-docker run --name=jersey-webservice-template -it -p 8080:8080 jack20191124/jersey-webservice-template
+docker run --name=fast-ws -it -p 8080:8080 jack20191124/fast-ws
 ```
 
-- __name=jersey-webservice-template__: the container is named "jersey-webservice-template". We can change it
+- __name=fast-ws__: the container is named "fast-ws". We can change it
   accordingly.
 - __-p 8080:8080__: 8080 is the port where webservice will listen on. With this port forwarding, we will be able to
   access webservice from host machine web browser at `localhost:8080`
@@ -176,7 +175,7 @@ Running Webservice in Standalone Jetty
 
 ### Download Jetty
 
-For JDK __17__, which is the version JWT runs on, it's been tested that Jetty _11.0.15_ worked. Hence, we will use
+For JDK __17__, which is the version FastWS runs on, it's been tested that Jetty _11.0.15_ worked. Hence, we will use
 ["11.0.15" release](https://repo1.maven.org/maven2/org/eclipse/jetty/jetty-home/11.0.15/jetty-home-11.0.15.tar.gz) as
 an example:
 
@@ -253,12 +252,12 @@ If tabs still come out at 2 spaces when hitting TAB or Enter, not 4 spaces, try:
 2. Do you have any .editorconfig files anywhere in the path of that file? Settings from .editorconfig
    ("Settings | Editor | Code Style") have priority (will overwrite) over your IDE settings.
 
-[Docker]: https://github.com/QubitPi/jersey-webservice-template/blob/master/Dockerfile
-[Docker-based integration tests]: https://github.com/QubitPi/jersey-webservice-template/blob/master/src/test/groovy/com/qubitpi/ws/jersey/template/DockerITSpec.groovy
+[Docker]: https://github.com/QubitPi/fast-ws/blob/master/Dockerfile
+[Docker-based integration tests]: https://github.com/QubitPi/fast-ws/blob/master/src/test/groovy/com/qubitpi/ws/jersey/template/DockerITSpec.groovy
 
 [jcabi-mysql]: https://mysql.jcabi.com/
-[jersey-webservice-template Dockerfile]: https://github.com/QubitPi/jersey-webservice-template/blob/master/Dockerfile
+[fast-ws Dockerfile]: https://github.com/QubitPi/fast-ws/blob/master/Dockerfile
 
-[style config]: https://github.com/QubitPi/jersey-webservice-template/blob/master/Jersey-Webservice-Template-Project-intellij-code-style.xml
+[style config]: https://github.com/QubitPi/fast-ws/blob/master/FastWS-Project-intellij-code-style.xml
 
 [Testcontainers]: https://qubitpi.github.io/testcontainers-java/
