+++ docs/src/css/custom.css
@@ -21,7 +21,7 @@
 
 /* You can override the default Infima variables here. */
 :root {
-  --ifm-color-primary: #F7941E;
+  --ifm-color-primary: #F58D4E;
   --ifm-color-primary-dark: #29784c;
   --ifm-color-primary-darker: #277148;
   --ifm-color-primary-darkest: #205d3b;
