+++ src/main/java/com/qubitpi/fastws/config/ApplicationConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.qubitpi.ws.jersey.template.config;
+package com.qubitpi.fastws.config;
 
 import org.aeonbits.owner.Config;
